separate headers into sources
