+++ savannah/actor.hpp
@@ -1,80 +1,17 @@
 #ifndef SAVANNAH_ACTOR_HPP
 #define SAVANNAH_ACTOR_HPP
-#include"init.hpp"
 #include"plant.hpp"
 #include"herbivore.hpp"
 #include"carnivore.hpp"
 #include<vector>
 
-bool isInTheWindow(const Vec2& sub_coord) {
-    return (sub_coord.x > 0 && sub_coord.x < field_width&& sub_coord.y>0 && sub_coord.y < field_height);
-}
-bool plantIsCover(std::vector<Plant>& plant, const Vec2& born_coord) {
-    for (size_t j = 0; j < plant.size(); j++) {
-        if (plant[j].isCover(born_coord)) return true;
-    }
-    return false;
-}
-void plantBorn(std::vector<Plant>& plant, const Vec2& born_coord) {
-    if (!isInTheWindow(born_coord)) return;
-    if(plantIsCover(plant, born_coord)) return;
-    plant.emplace_back();
-    plant[plant.size() - 1].setCoord(born_coord);
-
-}
+bool isInTheWindow(const Vec2& sub_coord);
+bool plantIsCover(std::vector<Plant>& plant, const Vec2& born_coord);
+void plantBorn(std::vector<Plant>& plant, const Vec2& born_coord);
 //植物の繁殖
-void plantBreed(std::vector<Plant>& plant, const size_t i) {
-    //植物の座標
-    const Vec2 plant_c = plant[i].getCoord();
-    plantBorn(plant, Vec2(plant_c.x - 10, plant_c.y));
-    plantBorn(plant, Vec2(plant_c.x + 10, plant_c.y));
-    plantBorn(plant, Vec2(plant_c.x, plant_c.y - 10));
-    plantBorn(plant, Vec2(plant_c.x, plant_c.y + 10));
-}
-
+void plantBreed(std::vector<Plant>& plant, const size_t i);
 //植物の行動
-void plantBehavior(std::vector<Plant>& plant, const double mi_spf) {
-    //植物それぞれの生命活動
-    for (size_t i = 0; i < plant.size();) {
-        bool plant_is_die, plant_is_breed;
-        plant[i].lifeActivity(mi_spf, plant_is_die, plant_is_breed);
-        //植物の繁殖
-        if (plant_is_breed) plantBreed(plant, i);
-        //植物が寿命で死ぬ
-        if (plant_is_die) plant.erase(plant.begin() + i);
-        else i++;
-    }
-}
-
-void herbivoreBehavior(std::vector<Herbivore>& herbivore, std::vector<Plant>& plant, const double mi_spf) {
-    //草食動物それぞれの行動
-    for (size_t i = 0; i < herbivore.size();) {
-        auto& h = herbivore[i];
-        bool herbivore_is_die;
-        h.lifeActivity(mi_spf, herbivore_is_die);
-        if (herbivore_is_die) {
-            herbivore.erase(herbivore.begin() + i);
-        }
-        else {
-            herbivore[i].behavior(herbivore,plant,i);
-            i++;
-        }
-    }
-}
-
-void carnivoreBehavior(std::vector<Carnivore>& carnivore, std::vector<Herbivore>& herbivore, const double mi_spf) {
-    //肉食動物それぞれの行動
-    for (size_t i = 0; i < carnivore.size();) {
-        auto& c = carnivore[i];
-        bool carnivore_is_die;
-        c.lifeActivity(mi_spf, carnivore_is_die);
-        if (carnivore_is_die) {
-            carnivore.erase(carnivore.begin() + i);
-        }
-        else {
-            carnivore[i].behavior(carnivore, herbivore, i);
-            i++;
-        }
-    }
-}
+void plantBehavior(std::vector<Plant>& plant, const double mi_spf);
+void herbivoreBehavior(std::vector<Herbivore>& herbivore, std::vector<Plant>& plant, const double mi_spf);
+void carnivoreBehavior(std::vector<Carnivore>& carnivore, std::vector<Herbivore>& herbivore, const double mi_spf);
 #endif // !SAVANNAH_ACTOR_HPP
