+++ savannah/draw.hpp
@@ -1,0 +1,10 @@
+#ifndef SAVANNAH_DRAW_HPP
+#define SAVANNAH_DRAW_HPP
+
+#include "herbivore.hpp"
+#include "carnivore.hpp"
+#include "plant.hpp"
+#include "property.hpp"
+#include <DxLib.h>
+
+#endif // !SAVANNAH_DRAW_HPP
