+++ savannah/plant.hpp
@@ -20,27 +20,9 @@ public:
     bool isCover(const Vec2& sub_coord);
     //繁殖判定
     bool isBreed();
+
+    void draw(int camera_x, const int camera_y, const double camera_exrate)const;
 };
 
 
-//植物クラス実装
-void Plant::lifeActivity(const double mi_spf, bool& is_die, bool& is_breed) {
-    setSpf(mi_spf);
-    getOld();
-    breed += getSpf();
-    is_breed = isBreed();
-    is_die = (getAge() > one_year * lifespan);
-    Vec2 temp_coord = getCoord();
-    is_die = (temp_coord.x<0 || temp_coord.x>field_width || temp_coord.y<0 || temp_coord.y>field_height);
-}
-bool Plant::isCover(const Vec2& sub_coord)  {
-    return (getCoord().distance(sub_coord) < 100.0);
-}
-bool Plant::isBreed() {
-    if (breed > one_year && getAge() > one_year / 2) {
-        breed = 0.0;
-        return true;
-    }
-    return false;
-}
 #endif // !SAVANNAH_PLANT_HPP
