+++ savannah/property.hpp
@@ -1,0 +1,12 @@
+#ifndef SAVANNAH_PROPERTY_HPP
+#define SAVANNAH_PROPERTY_HPP
+
+extern int zebra1;
+extern int zebra2;
+extern int zebra3;
+extern int lion1;
+extern int lion2;
+extern int lion3;
+extern int grass1;
+
+#endif // !SAVANNAH_PROPERTY_HPP
